move vc related modules into vc folder
