+++ src/vc/configs/32k.json
