+++ src/vc/infer_pack/attentions.py
@@ -5,9 +5,9 @@ import torch
 from torch import nn
 from torch.nn import functional as F
 
-from infer_pack import commons
-from infer_pack import modules
-from infer_pack.modules import LayerNorm
+from vc.infer_pack import commons
+from vc.infer_pack import modules
+from vc.infer_pack.modules import LayerNorm
 
 
 class Encoder(nn.Module):
