+++ src/vc/infer_pack/models_onnx.py
@@ -3,15 +3,15 @@ from time import time as ttime
 import torch
 from torch import nn
 from torch.nn import functional as F
-from infer_pack import modules
-from infer_pack import attentions
-from infer_pack import commons
-from infer_pack.commons import init_weights, get_padding
+from vc.infer_pack import modules
+from vc.infer_pack import attentions
+from vc.infer_pack import commons
+from vc.infer_pack.commons import init_weights, get_padding
 from torch.nn import Conv1d, ConvTranspose1d, AvgPool1d, Conv2d
 from torch.nn.utils import weight_norm, remove_weight_norm, spectral_norm
-from infer_pack.commons import init_weights
+from vc.infer_pack.commons import init_weights
 import numpy as np
-from infer_pack import commons
+from vc.infer_pack import commons
 
 
 class TextEncoder256(nn.Module):
