+++ src/vc/infer_pack/modules.py
@@ -9,9 +9,9 @@ from torch.nn import functional as F
 from torch.nn import Conv1d, ConvTranspose1d, AvgPool1d, Conv2d
 from torch.nn.utils import weight_norm, remove_weight_norm
 
-from infer_pack import commons
-from infer_pack.commons import init_weights, get_padding
-from infer_pack.transforms import piecewise_rational_quadratic_transform
+from vc.infer_pack import commons
+from vc.infer_pack.commons import init_weights, get_padding
+from vc.infer_pack.transforms import piecewise_rational_quadratic_transform
 
 
 LRELU_SLOPE = 0.1
